+++ packages/shared-fixtures/fixtures/typescript/basics/optional-chain-call.src.ts
@@ -1,0 +1,12 @@
+function processOptionalCall(one?: any) {
+  one?.fn();
+  one?.two.fn();
+  one.two?.fn();
+  one.two?.three.fn();
+  one.two?.three?.fn();
+
+  one?.();
+  one?.()?.();
+
+  one?.().two;
+}